refactor scala code structure
