+++ zoo/src/main/scala/com/intel/analytics/zoo/common/Utils.scala
@@ -20,17 +20,45 @@ import java.io._
 import java.nio.file.attribute.PosixFilePermissions
 import java.nio.file.{Path => JPath}
 
+import com.intel.analytics.bigdl.nn.abstractnn.Activity
+import com.intel.analytics.bigdl.tensor.Tensor
 import org.apache.commons.io.filefilter.WildcardFileFilter
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.{FSDataInputStream, FSDataOutputStream, FileSystem, Path}
 import org.apache.hadoop.io.IOUtils
 import org.apache.log4j.Logger
 
+import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
 
 private[zoo] object Utils {
 
   private val logger = Logger.getLogger(getClass)
+
+  @inline
+  def timeIt[T](name: String)(f: => T): T = {
+    val begin = System.nanoTime()
+    val result = f
+    val end = System.nanoTime()
+    val cost = end - begin
+    logger.debug(s"$name time [${cost / 1.0e9} s].")
+    result
+  }
+
+  def activity2VectorBuilder(data: Activity):
+  mutable.Builder[Tensor[_], Vector[Tensor[_]]] = {
+    val vec = Vector.newBuilder[Tensor[_]]
+    if (data.isTensor) {
+      vec += data.asInstanceOf[Tensor[_]]
+    } else {
+      var i = 0
+      while (i < data.toTable.length()) {
+        vec += data.toTable(i + 1)
+        i += 1
+      }
+    }
+    vec
+  }
 
   def listLocalFiles(path: String): Array[File] = {
     val files = new ArrayBuffer[File]()
