+++ zoo/src/main/scala/com/intel/analytics/zoo/tfpark/TFTensorNumeric.scala
@@ -13,12 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.analytics.zoo.pipeline.api.net
+package com.intel.analytics.zoo.tfpark
 
-
-
-import com.intel.analytics.bigdl.tensor.{ConvertableFrom, StringType, TensorDataType}
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.UndefinedTensorNumeric
+import com.intel.analytics.bigdl.tensor.{ConvertableFrom, StringType, TensorDataType}
 
 import scala.language.implicitConversions
 
@@ -34,4 +32,3 @@ object TFTensorNumeric {
 
   }
 }
-
